requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
aeppl=0.0.38=pyhd8ed1ab_0
aeppl-base=0.0.38=pyhd8ed1ab_0
aesara=2.8.7=py310h18ae672_0
aesara-base=2.8.7=py310hff52083_0
arviz=0.13.0=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
binutils=2.39=hdd6e379_0
binutils_impl_linux-64=2.39=h6ceecb4_0
binutils_linux-64=2.39=h5fc0e48_11
blas=2.116=openblas
blas-devel=3.9.0=16_linux64_openblas
blosc=1.21.1=h83bc5f7_3
boost-cpp=1.78.0=h75c5d50_1
branca=0.5.0=pyhd8ed1ab_0
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py310h5764c6d_1005
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
c-compiler=1.5.1=h166bdaf_0
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
cairo=1.16.0=ha61ee94_1014
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310h255011f_2
cfitsio=4.1.0=hd9d235c_0
cftime=1.6.2=py310hde88566_1
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
cloudpickle=2.2.0=pyhd8ed1ab_0
colorlover=0.3.0=py_0
cons=0.4.5=pyhd8ed1ab_0
contourpy=1.0.5=py310hbf28c38_1
cryptography=38.0.2=py310h600f1e7_2
curl=7.86.0=h2283fc2_0
cxx-compiler=1.5.1=h924138e_0
cycler=0.11.0=pyhd8ed1ab_0
etuples=0.3.8=pyhd8ed1ab_0
expat=2.5.0=h27087fc_0
fastprogress=1.0.3=pyhd8ed1ab_0
filelock=3.8.0=pyhd8ed1ab_0
fiona=1.8.22=py310h60a68a4_0
folium=0.13.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.1=hc2a2eb6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=py310h5764c6d_1
freetype=2.12.1=hca18f0e_0
freexl=1.0.6=h166bdaf_1
gcc=10.4.0=hb92f740_11
gcc_impl_linux-64=10.4.0=h5231bdf_19
gcc_linux-64=10.4.0=h9215b83_11
gdal=3.5.2=py310he53f9b6_5
geopandas=0.12.0=pyhd8ed1ab_0
geopandas-base=0.12.0=pyha770c72_0
geos=3.11.0=h27087fc_0
geotiff=1.7.1=ha76d385_4
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h36c2ea0_2
gxx=10.4.0=hb92f740_11
gxx_impl_linux-64=10.4.0=h5231bdf_19
gxx_linux-64=10.4.0=h6e491c6_11
hdf4=4.2.15=h9772cbc_4
hdf5=1.12.2=nompi_h4df4325_100
icu=70.1=h27087fc_0
idna=3.4=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h166bdaf_2
json-c=0.16=hc379101_0
kealib=1.4.15=ha7026e8_1
kernel-headers_linux-64=2.6.32=he073ed8_15
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py310hbf28c38_1
krb5=1.19.3=h08a2579_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.39=hc81fddc_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=16_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libcblas=3.9.0=16_linux64_openblas
libcurl=7.86.0=h2283fc2_0
libdap4=3.20.6=hd7c4107_2
libdeflate=1.14=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h7f98852_5
libgcc-devel_linux-64=10.4.0=hd38fd1e_19
libgcc-ng=12.2.0=h65d4601_19
libgdal=3.5.2=h27ae5c1_5
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.74.1=h7a41b64_0
libgomp=12.2.0=h65d4601_19
libiconv=1.17=h166bdaf_0
libkml=1.3.0=h37653c0_1015
liblapack=3.9.0=16_linux64_openblas
liblapacke=3.9.0=16_linux64_openblas
libnetcdf=4.8.1=nompi_h21705cb_104
libnghttp2=1.47.0=hff17c54_1
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=pthreads_h78a6416_3
libpng=1.6.38=h753d276_0
libpq=14.5=he2d8382_1
librttopo=1.1.0=hf730bdb_11
libsanitizer=10.4.0=h5246dfb_19
libspatialindex=1.9.3=h9c3ff4c_4
libspatialite=5.0.1=hfbd986c_21
libsqlite=3.39.4=h753d276_0
libssh2=1.10.0=hf14f497_3
libstdcxx-devel_linux-64=10.4.0=hd38fd1e_19
libstdcxx-ng=12.2.0=h46fd767_19
libtiff=4.4.0=h55922b4_4
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.4=h166bdaf_0
libxcb=1.13=h7f98852_1004
libxml2=2.10.3=h7463322_0
libzip=1.9.2=hc929e4a_1
libzlib=1.2.13=h166bdaf_4
llvm-openmp=14.0.4=he0ac6c6_0
logical-unification=0.4.5=pyhd8ed1ab_0
lz4-c=1.9.3=h9c3ff4c_1
mapclassify=2.4.3=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_2
matplotlib-base=3.6.1=py310h8d5ebf3_0
minikanren=1.0.3=pyhd8ed1ab_0
mkl=2022.1.0=h84fe81f_915
mkl-service=2.4.0=py310h9263142_0
multipledispatch=0.6.0=py_0
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h27087fc_1
netcdf4=1.6.1=nompi_py310h55e1e36_101
networkx=2.8.7=pyhd8ed1ab_0
nspr=4.32=h9c3ff4c_1
nss=3.78=h2350873_0
numpy=1.23.4=py310h53a5b5f_1
openblas=0.3.21=pthreads_h320a7e8_3
openjpeg=2.5.0=h7d73246_1
openssl=3.0.5=h166bdaf_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.1=py310h769672d_1
pcre=8.45=h9c3ff4c_0
pcre2=10.37=hc3806b6_1
pillow=9.2.0=py310hbd86126_2
pip=22.3=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
poppler=22.10.0=h92391eb_0
poppler-data=0.4.11=hd8ed1ab_0
postgresql=14.5=ha105346_1
proj=9.1.0=h93bde94_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pygeos=0.13=py310h5e49deb_0
pymc=4.2.2=hd8ed1ab_0
pymc-base=4.2.2=pyhd8ed1ab_0
pyopenssl=22.1.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyproj=3.4.0=py310hb1338dc_1
pysocks=1.7.1=pyha2e5f31_6
python=3.10.6=ha86cf86_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.5=pyhd8ed1ab_0
readline=8.1.2=h0f457ee_0
requests=2.28.1=pyhd8ed1ab_1
rtree=1.0.1=py310hbdcdc62_1
scikit-learn=1.1.3=py310h0c3af53_0
scipy=1.9.3=py310hdfbd76f_1
setuptools=65.5.0=pyhd8ed1ab_0
shapely=1.8.5=py310h5e49deb_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=hbd366e4_1
sqlite=3.39.4=h4ff8645_0
sysroot_linux-64=2.12=he073ed8_15
tbb=2021.6.0=h924138e_1
threadpoolctl=3.1.0=pyh8a188c0_0
tiledb=2.11.3=h3f4058f_1
tk=8.6.12=h27826a3_0
toolz=0.12.0=pyhd8ed1ab_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzcode=2022e=h166bdaf_0
tzdata=2022e=h191b570_0
unicodedata2=14.0.0=py310h5764c6d_2
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xarray=2022.10.0=pyhd8ed1ab_0
xarray-einstats=0.3.0=pyhd8ed1ab_0
xerces-c=3.2.4=h55805fa_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xyzservices=2022.9.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h6239696_4
